Standalone Inventory System
Have a basic standalone inventory system up and running. It is limited currently in that it only works with itself, but it has the ability to store and maintain items, items are movable within the inventory, and you can left click to drag and drop an item anywhere and right click to place a single item of the stack. Stacks will automatically attempt to merge if possible.