Add and Drop Items
Added functionality to add and drop items. To add items you instantiate the pick up scene located in the Items folder, set its' data to the item you want it to be, and establish it's position in the world. It will automatically look for existing items in the inventory and add to them if possible, otherwise it will add to the first blank slot until the inventory is filled. Items can be dropped from grabbing an item in your inventory and clicking somewhere outside of your inventory. Left click drops the stack, right click drops a single item. Additionally, if you are holding an item and you close the inventory the item stack will automatically be dropped.